# expression: implement vectorized evaluation for `builtinArithmeticModDecimalSig`