### What problem does this PR solve?

Implement vectorized evaluation for builtinArithmeticModDecimalSig.#12102

### What is changed and how it works?
```
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticModDecimalSig-VecBuiltinFunc-4                            4634            281010 ns/op   30000 B/op        628 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticModDecimalSig-NonVecBuiltinFunc-4                         2334            517156 ns/op   60144 B/op       1256 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticModDecimalSig-VecBuiltinFunc#01-4                        10000            109277 ns/op       0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticModDecimalSig-NonVecBuiltinFunc#01-4                      3506            306647 ns/op   39888 B/op        831 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticModDecimalSig-VecBuiltinFunc#02-4                        15810             83592 ns/op       0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticModDecimalSig-NonVecBuiltinFunc#02-4                      5018            238846 ns/op   38976 B/op        812 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticModDecimalSig-VecBuiltinFunc#03-4                        92685             12641 ns/op       0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticModDecimalSig-NonVecBuiltinFunc#03-4                     14545             69335 ns/op       0 B/op          0 allocs/op
```

### Check List

Tests
- Unit test